mic-tester: pull the media api stuff into a service.
